test(phd2-guider): consolidate integration test binaries (#163)
Phase 4 of build-optimization-test-reorganization. Merge
test_main_integration.rs into test_integration.rs to eliminate one
test binary; both spawn processes and test end-to-end behavior.
test_mock_server.rs stays separate (TCP binding + client internals).

- 3 test files -> 2 test files
- 38 CLI subprocess tests appended under "CLI Subprocess Tests" section
- Deduped get_available_port helper
- Added Output to std::process import
- Updated docs/services/phd2-guider.md

Verified: cargo nextest run -p phd2-guider passes 344/344 (10
ignored), cargo rail run --profile commit clean, cargo fmt clean.
